fix(audio): reset the decoder on seek and surface failed reads
Seeking resets the decoder before refining the position. Symphonia requires
this whenever the next packet is discontinuous with the last decoded one;
for MP3 the state carried across is the bit reservoir and the synthesis
overlap, so without it residual audio from before the seek bleeds into the
frames after it while the frame count stays correct.

Initialization no longer treats a failed packet read as end of media. Only a
clean `Ok(None)` ends a stream, and reaching it before any packet decoded to
audio is an error rather than a zero-length track: with gapless trimming the
first packet can legitimately decode to nothing, so an empty buffer is no
longer evidence that the stream finished.

The seek retry's frame arithmetic moves into `refined_offset_frames`, which
can be tested without driving a decode failure through the demuxer, and the
fixture-backed tests move into a `#[path]` child module so the decoder logic
stays navigable.
